// types/companies.ts
export type CompetitorsResultsType = {
  name: string;
  tag?: string;
  number_of_reviews: number;
  average_rating: number;
  industry: string;
  rank: number;
  url?: string;
};